feat: extract password strength component for reusability
- Created new PasswordStrength component in `passwordStrength.tsx`
- Refactored password strength UI from RegisterStep6 into a standalone component
- Maintained existing functionality with improved component modularity
- Added props interface for flexible usage across different forms
